% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circle_pack.R
\name{compute_panel_circle_pack}
\alias{compute_panel_circle_pack}
\title{compute_panel_circle_pack}
\usage{
compute_panel_circle_pack(data, scales)
}
\arguments{
\item{scales}{}
}
\description{
compute_panel_circle_pack
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
gapminder::gapminder \%>\%
filter(continent == "Americas") \%>\%
  filter(year == 2002) \%>\%
  # input must have required aesthetic inputs as columns
  rename(area = pop) \%>\%
  compute_panel_circle_pack() \%>\%
  head()

  gapminder::gapminder \%>\%
filter(continent == "Americas") \%>\%
 filter(year == 2002) \%>\%
 # input must have required aesthetic inputs as columns
 rename(area = pop) \%>\%
 compute_panel_circle_pack() \%>\%
 ggplot() +
 aes(x = x, y = y, fill = country) +
 geom_polygon()
}
