% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circle_pack_collapse.R
\name{compute_panel_circle_pack_collapse}
\alias{compute_panel_circle_pack_collapse}
\title{compute_panel_circle_pack}
\usage{
compute_panel_circle_pack_collapse(data, scales, fun = sum)
}
\arguments{
\item{scales}{}
}
\description{
compute_panel_circle_pack
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
gapminder::gapminder \%>\%
group_by(continent) \%>\%
  filter(year == 2002) \%>\%
  # input must have required aesthetic inputs as columns
  count() \%>\%
  rename(area = n)

  gapminder::gapminder \%>\%
filter(continent == "Americas") \%>\%
 filter(year == 2002) \%>\%
 # input must have required aesthetic inputs as columns
 rename(area = pop) \%>\%
 compute_panel_circle_pack() \%>\%
 ggplot() +
 aes(x = x, y = y, fill = country) +
 geom_polygon()
}
