% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_panel_circlepack.R
\name{compute_panel_circlepack}
\alias{compute_panel_circlepack}
\title{compute_panel_circlepack}
\usage{
compute_panel_circlepack(data, scales)
}
\arguments{
\item{scales}{}
}
\description{
compute_panel_circlepack
}
\examples{
TBD
}
