% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circle_pack_collapse.R
\name{geom_polygon_circlepack_collapse}
\alias{geom_polygon_circlepack_collapse}
\title{Title}
\usage{
geom_polygon_circlepack_collapse(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
gapminder::gapminder \%>\%
filter(year == 2002) \%>\%
  ggplot() +
  aes(id = continent) +
  geom_polygon_circlepack_collapse(alpha = .5, size = .002)

last_plot() +
  aes(fill = continent)

last_plot() +
  aes(area = pop)

last_plot() +
  aes(color = continent) +
  facet_wrap(facets = vars(continent))
}
