% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_circlepack_text.R
\name{stat_circlepack_center}
\alias{stat_circlepack_center}
\title{Title}
\usage{
stat_circlepack_center(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
