% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_panel_circlepack_center.R
\name{compute_panel_circlepack_center}
\alias{compute_panel_circlepack_center}
\title{compute_panel_circlepack_center}
\usage{
compute_panel_circlepack_center(data, scales)
}
\description{
compute_panel_circlepack_center
}
